requirements.txt
# Core dependencies
sqlalchemy==2.0.23
alembic==1.13.1
python-dotenv==1.0.0
tabulate==0.9.0
colorama==0.4.6

# Email functionality
secure-smtplib==0.1.1

# Database
psycopg2-binary==2.9.9

# Utilities
python-dateutil==2.8.2
